requirements/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
mysql-connector-python
python-dotenv
kiteconnect
holidays